+++ CHANGELOG.md
@@ -7,16 +7,17 @@ Beta Version
 0.2 All core functions included
 0.3 Fixed dependencies when installing in fresh environments
 0.3.1 Fixed a typo in dependencies
-0.3.2 Updated functionality of seabass (automatic sleep stage detection)
+0.3.2 Added functionality of seabass (automatic sleep stage detection)
 0.3.2 Bug fixes in stage and spindle detection for continuity when some subs crash
 0.3.3 Fixed seabass when some channels (e.g. emg, eog) missing
 0.3.4 More bug fixes in spindle detection for when running only some subjects at a time
 0.3.5 Added functionality to read sub and ses from tracking sheet
 0.3.7 Removed redundant scripts from pac
-0.3.9 Added artefact detection 
+0.3.9 Added first artefact detection 
 0.4 Added continuous PAC for and fixed bugs with event PAC
 0.4.1 Updated format bids to allow subject-level
 0.4.2 Updates to export PAC summary
 0.4.3 Improved loading channel names and completed PAC event dataset
 0.4.4 Fixed bugs related to reading channel names from tracking sheet
 0.4.5 Fixed bugs with power spectrum analysis and export
+0.4.6 Fixed bugs with audit, whales, logs and added first REMS detection (YASA)